AdaptiveSparseGrids

AdaptiveSparseGrids.drive_to_college!

AdaptiveSparseGrids.refinegrid!

AdaptiveSparseGrids.drive_to_college! — Method

This function inserts the children into the list of nodes, and sets up the parent/child linkages that allow the child to be used in function evaluations

AdaptiveSparseGrids.refinegrid! — Method

Proceeds in 4 steps 1) Obtain the list of possible child nodes to be created 2) Remove duplicates 3) Evalute the function 4) Insert the points into the grid